refactor: apply trade type tabs to vanilla
