+++ packages/shared/src/utils/constants/contract.tsx
@@ -155,7 +155,7 @@ export const getContractTypesConfig: TGetContractTypesConfig = symbol => ({
         title: localize('Call/Put'),
         trade_types: ['VANILLALONGCALL', 'VANILLALONGPUT'],
         basis: ['stake'],
-        components: ['duration', 'strike', 'amount', 'vanilla_trade_type'],
+        components: ['duration', 'strike', 'amount', 'trade_type_tabs'],
         barrier_count: 1,
         config: { should_override: true },
     },
