+++ packages/trader/src/Modules/Trading/Components/Form/screen-small.jsx
@@ -19,9 +19,8 @@ import Purchase from '../../Containers/purchase.jsx';
 import 'Sass/app/_common/mobile-widget.scss';
 import classNames from 'classnames';
 import PayoutPerPointMobile from '../Elements/payout-per-point-mobile';
-import TradeTypeTabs from './TradeParams/Turbos/trade-type-tabs';
+import TradeTypeTabs from './TradeParams/trade-type-tabs';
 import Strike from 'Modules/Trading/Components/Form/TradeParams/strike.jsx';
-import VanillaTradeTypes from 'Modules/Trading/Components/Form/TradeParams/vanilla-trade-types.jsx';
 
 const CollapsibleTradeParams = ({
     form_components,
@@ -57,7 +56,6 @@ const CollapsibleTradeParams = ({
                 <ContractType />
                 {is_multiplier && <MultiplierOptionsWidget />}
                 {isVisible('trade_type_tabs') && <TradeTypeTabs />}
-                {is_vanilla && <VanillaTradeTypes />}
             </div>
             {isVisible('last_digit') && (
                 <div collapsible='true'>
