+++ packages/trader/src/Modules/Trading/Containers/trade-params.jsx
@@ -12,7 +12,7 @@ import StopLoss from 'Modules/Trading/Components/Form/TradeParams/Multiplier/sto
 import TakeProfit from 'Modules/Trading/Components/Form/TradeParams/Multiplier/take-profit.jsx';
 import Expiration from 'Modules/Trading/Components/Form/TradeParams/Multiplier/expiration.jsx';
 import Strike from 'Modules/Trading/Components/Form/TradeParams/strike.jsx';
-import VanillaTradeTypes from 'Modules/Trading/Components/Form/TradeParams/vanilla-trade-types.jsx';
+import TradeTypeTabs from 'Modules/Trading/Components/Form/TradeParams/trade-type-tabs';
 import { connect } from 'Stores/connect';
 import Fieldset from 'App/Components/Form/fieldset.jsx';
 
@@ -25,12 +25,12 @@ const TradeParams = ({ form_components, is_minimized }) => {
         <React.Fragment>
             {isVisible('duration') && <Duration key={'duration'} is_minimized={is_minimized} />}
             {isVisible('barrier') && <Barrier key={'barrier'} is_minimized={is_minimized} />}
-            {isVisible('barrier_selector') && <BarrierSelector key={'barrier_selector'} />}
             {isVisible('last_digit') && <LastDigit key={'last_digit'} is_minimized={is_minimized} />}
-            {(isVisible('vanilla_trade_type') || isVisible('strike')) && (
+            {(isVisible('trade_type_tabs') || isVisible('strike') || isVisible('barrier_selector')) && (
                 <Fieldset className={classNames('trade-container__fieldset', 'trade-container__fieldset--no-padding')}>
-                    {isVisible('vanilla_trade_type') && <VanillaTradeTypes key={'vanilla_trade_type'} />}
+                    {isVisible('trade_type_tabs') && <TradeTypeTabs key={'trade_type_tabs'} />}
                     {isVisible('strike') && <Strike key={'strike'} />}
+                    {isVisible('barrier_selector') && <BarrierSelector key={'barrier_selector'} />}
                 </Fieldset>
             )}
             {isVisible('amount') && <Amount key={'amount'} is_minimized={is_minimized} />}
